==> Src/01 - Dataset Cleaning.Rmd <==
---
title: "NPS Dataset Cleaning"
output: html_document
---

Import data:
```{r}
library(readxl)
data <- read_excel("../Dataset/UCS-Satellite-Database-1-1-2023.xlsx", col_names = TRUE)
```


PULIZIA INIZIALE: 


Nomi colonne:
```{r}
names(data)
```

Rimozione colonne inutili:
```{r}
columns_to_remove <- c("Name of Satellite, Alternate Names", "Country/Org of UN Registry", "Operator/Owner", "Detailed Purpose", "Type of Orbit", "Contractor", "Launch Site", "Launch Vehicle", "COSPAR Number", "NORAD Number", "Comments", "...32", "Source Used for Orbital Data", "Source...34", "Source...35", "Source...36", "Source...37", "Source...38", "Source...39", "Source...40")

data <- data[, -c(which(names(data) %in% columns_to_remove), 41:71)]

names(data)
```

Confronta "Country of Contractor" con "Country of Operator/Owner":
```{r}
ind <- which(data$`Country of Contractor` != data$`Country of Operator/Owner`)

cc <- data$`Country of Contractor`[ind]
co <- data$`Country of Operator/Owner`[ind]

df <- data.frame(co, cc)
df
```

1139 diversi, teniamo country of operator:
```{r}
data <- data[, -which(names(data) == "Country of Contractor")]
```

Summary:
```{r}
summary(data)
```


Rendi numeriche le colonne Dry mass e Longitude of GEO (sono character)
```{r}
data$`Dry Mass (kg.)` <- as.numeric(data$`Dry Mass (kg.)`)
data$`Longitude of GEO (degrees)` <- as.numeric(data$`Longitude of GEO (degrees)`)
```



Numero di NA in ogni colonna
```{r}
df <- data.frame()

for (colonna in colnames(data)) {
    conteggio_na <- sum(is.na(data[[colonna]]))
    df <- rbind(df, data.frame(Colonna = colonna, Numero_di_NA = conteggio_na))
}

df
```

Togliamo colonne launch mass(perchè teniamo dry mass) e power (6137 NA)
```{r}
columns_to_remove <- c("Power (watts)", "Launch Mass (kg.)")

data <- data[, -which(names(data) %in% columns_to_remove)]
```



Expected Lifetime (yrs.) ha 1915 NA, toglierei le righe senza questo dato
Date of Launch ha 2 NA, toglierei le righe senza questo dato
Dry Mass (kg.) ha 6278 NA, toglierei le righe senza questo dato
Period ha 45 NA, toglierei le righe senza questo dato
Inclination (degrees) ha 3 NA, toglierei le righe senza questo dato
Eccentricity ha 1 NA, toglierei le righe senza questo dato
Apogee ha 1 NA, toglierei le righe senza questo dato
Perigree ha 1 NA, toglierei le righe senza questo dato
Longitude of GEO ha 3 NA, toglierei le righe senza questo dato

```{r}
cols <- c("Expected Lifetime (yrs.)", "Date of Launch", "Dry Mass (kg.)", "Period (minutes)", "Inclination (degrees)", "Eccentricity", "Apogee (km)", "Perigee (km)", "Longitude of GEO (degrees)")

data <- data[complete.cases(data[, cols]), ]
```


Numero di NA in ogni colonna
```{r}
df <- data.frame()

for (colonna in colnames(data)) {
    conteggio_na <- sum(is.na(data[[colonna]]))
    df <- rbind(df, data.frame(Colonna = colonna, Numero_di_NA = conteggio_na))
}

df
```

Colonne sopravvissute:
```{r}
names(data)
```

Sistemare Nomi:

1) Stati

ESA/Russia -> ESA
Turkmenistan/Monaco -> Turkmenistan
United Kingdom/ESA -> United Kingdom
USA/Canada -> USA
USA/France -> USA
USA/Japan -> USA
USA/Japan/Brazil -> USA
```{r}
data$'Country of Operator/Owner'[which(data$'Country of Operator/Owner' == "ESA/Russia")] <- "ESA"
data$'Country of Operator/Owner'[which(data$'Country of Operator/Owner' == "Turkmenistan/Monaco")] <- "Turkmenistan"
data$'Country of Operator/Owner'[which(data$'Country of Operator/Owner' == "United Kingdom/ESA")] <- "United Kingdom"
data$'Country of Operator/Owner'[which(data$'Country of Operator/Owner' == "USA/Canada")] <- "USA"
data$'Country of Operator/Owner'[which(data$'Country of Operator/Owner' == "USA/France")] <- "USA"
data$'Country of Operator/Owner'[which(data$'Country of Operator/Owner' == "USA/Japan")] <- "USA"
data$'Country of Operator/Owner'[which(data$'Country of Operator/Owner' == "USA/Japan/Brazil")] <- "USA"
```


2) Users
"Government/Civil" -> "Government"
"Government/Commercial" -> "Government"
"Government/Commercial/Military" -> "Government"
"Military/Commercial" -> "Military"
"Military/Government" -> "Military"
```{r}
data$Users[which(data$Users == "Government/Civil")] <- "Government"
data$Users[which(data$Users == "Government/Commercial")] <- "Government"
data$Users[which(data$Users == "Government/Commercial/Military")] <- "Government"
data$Users[which(data$Users == "Military/Commercial")] <- "Military"
data$Users[which(data$Users == "Military/Government")] <- "Military"
```

3) Purpose
"Communications/Navigation" -> "Communications"
"Communications/Technology Development" -> "Communications"
"Earth Observation/Technology Development" -> "Earth Observation"
"Navigation/Global Positioning" -> "Navigation"
"Navigation/Regional Positioning" -> "Navigation"
```{r}
data$Purpose[which(data$Purpose == "Communications/Navigation")] <- "Communications"
data$Purpose[which(data$Purpose == "Communications/Technology Development")] <- "Communications"
data$Purpose[which(data$Purpose == "Earth Observation/Technology Development")] <- "Earth Observation"
data$Purpose[which(data$Purpose == "Navigation/Global Positioning")] <- "Navigation"
data$Purpose[which(data$Purpose == "Navigation/Regional Positioning")] <- "Navigation"
```

Rendi colonne factor:
```{r}
data$'Country of Operator/Owner' <- factor(data$'Country of Operator/Owner', ordered = F)
print('Country of Operator/Owner')
table(data$'Country of Operator/Owner')

data$Users <- factor(data$Users, ordered = F)
print('Users')
table(data$Users)

data$Purpose <- factor(data$Purpose, ordered = F)
print('Purpose')
table(data$Purpose)

data$'Class of Orbit' <- factor(data$'Class of Orbit', ordered = F)
print('Class of Orbit')
table(data$'Class of Orbit')

data$Status <- factor(data$Status, ordered = F)
print('Status')
table(data$Status)
```


Ispezione del dataset:
```{r}
summary(data)
```

Cambiamo i nomi delle colonne:

"Current Official Name of Satellite" -> "Satellite Name"
"Country of Operator/Owner" -> "Country"
"Users" -> "Users"
"Purpose" -> "Purpose"
"Class of Orbit" -> "Orbit" 
"Longitude of GEO (degrees)" -> "Longitude of GEO"
"Perigee (km)" -> "Perigee"
"Apogee (km)" -> "Apogee"
"Eccentricity" -> "Eccentricity"
"Inclination (degrees)" -> "Inclination"
"Period (minutes)" -> "Period"
"Dry Mass (kg.)" -> "Mass" 
"Date of Launch" -> "Launch Date"
"Status" -> "Status"
"Date of Disposal" -> "Final Date"
"Effective Lifetime (yrs.)" -> "Effective Lifetime"
"Expected Lifetime (yrs.)" -> "Expected Lifetime"

```{r}
names(data) <- c("Satellite Name", "Country", "Users", "Purpose", "Orbit", "Longitude of GEO", "Perigee", "Apogee", "Eccentricity", "Inclination", "Period", "Mass", "Launch Date", "Status", "Final Date", "Effective Lifetime", "Expected Lifetime")
```


Impostare come data finale 17/11/2023 se lo status è attivo:
```{r}
data$'Final Date'[which(data$Status == "active")] <- as.POSIXct("2023/11/17", "UTC")
```

Calcolo effective lifetime:
```{r}
data$`Effective Lifetime` <- round(as.numeric(data$'Final Date' - data$`Launch Date`)/365, digit = 2)
```

Esportazione dataset
```{r}
# Caricare la libreria
library(openxlsx)
# Specificare il percorso del file Excel
percorso_file_excel <- "C:/Users/simon/Desktop/Università/V Anno/I Semester/Nonparametric Statistics/Project"

# Scrivere il dataframe nel file Excel
write.xlsx(data, percorso_file_excel, rowNames = FALSE)
```


COSTRUZIONE JOIN DATA

Lettura datasets:
```{r}
library(readxl)
types = rep('guess',68)
types[20] = 'date'
dataUCS <- read_excel("../Dataset/UCS-Satellite-Database-1-1-2023.xlsx", col_names = TRUE, col_types = types)
dataSatCat <- read.csv("../Dataset/satcat.csv", header = TRUE)
```


```{r}
# Trova gli indici in cui gli elementi nei due vettori sono uguali
id1 <- which(datainiz$`COSPAR Number` %in% dataSatCat$OBJECT_ID)
id2 <- which(dataSatCat$OBJECT_ID %in% dataUCS$`COSPAR Number`)
```


```{r}
#prendo solo colonne OBJECT_ID e DECAY_DATE
dataSatCat <- dataSatCat[id2,c(2,9)]

#tolgo le osservazioni senza DECAY_DATE
id <- which(dataSatCat$DECAY_DATE == "")
dataSatCat <- dataSatCat[-id,]

#ricalcolo indice in cui gli elementi nei due vettori sono uguali
id1 <- which(dataUCS$`COSPAR Number` %in% dataSatCat$OBJECT_ID)
```

Tolgo colonne utili da dataset UCS
```{r}
columns_to_remove <- c("Name of Satellite, Alternate Names", "Country/Org of UN Registry", "Operator/Owner", "Detailed Purpose", "Type of Orbit", "Contractor", "Launch Site", "Launch Vehicle", "COSPAR Number", "NORAD Number", "Comments", "...32", "Source Used for Orbital Data", "Source...34", "Source...35", "Source...36", "Source...37", "Source...38", "Source...39", "Source...40")

dataUCS <- dataUCS[, -c(which(names(dataUCS) %in% columns_to_remove), 41:68)]
dataUCS <- dataUCS[, -c(17:24)]
```

Per i satelliti in comune inserisco nella Disposal Date le Decay_Date del satellite prese dal dataset SatCat
```{r}
dataUCS$`Disposal Date` <- vector("character", length = nrow(dataUCS))
dataUCS$`Disposal Date`[id1] <- dataSatCat$DECAY_DATE
```

Tolgo da dataset UCS i satelliti senza disposal date
```{r}
id <- which(dataUCS$`Disposal Date` == "")
# dataset con disposal date 
dataUCS <- dataUCS[-id,]
```

Adesso voglio unire i satelliti presenti in dataUCS con quelli che avevamo ottenuto dal dataset precedente.
Per fare ciò devo modificare le variabili in modo che siano uguali tra i due datasets. (Ad esempio devo recuperare la lauch mass che precedentemente 
avevamo tolto)
```{r}
# prendo dataset aggiustato precedentemente 
data_prec <- read_excel("../Dataset/Data_Satellites.xlsx", col_names = TRUE)

# controllo satelliti in comune ( per capire che satelliti devo aggiungere)
id <- which(dataUCS$`Current Official Name of Satellite` %in% data_prec$`Satellite Name`)

# aggiusto variabili
dataUCS$`Dry Mass (kg.)` <- as.numeric(dataUCS$`Dry Mass (kg.)`)
dataUCS$`Longitude of GEO (degrees)` <- as.numeric(dataUCS$`Longitude of GEO (degrees)`)
# tolgo dry mass e power
columns_to_remove <- c("Power (watts)", "Dry Mass (kg.)") 
dataUCS <- dataUCS[, -which(names(dataUCS) %in% columns_to_remove)]
#aggiungo status e effective lifetime
dataUCS$Status <- 'retired'
dataUCS$`Effective Lifetime` <- vector("character", length = nrow(dataUCS))
```


```{r}
# creo subset del dataset (satelliti dal sat da aggiungere a dataUCS)
sat.da.agg.a.dataUCS <- data_prec[-id,] 

# recupero Launch mass dei satelliti che non la hanno ( quelli che derivano dal nostro dataset modificato precedentemente)
id1 <- which(sat.da.agg.a.dataUCS$`Satellite Name` %in% data_prec$`Current Official Name of Satellite`)
data <- read_excel("../Dataset/UCS-Satellite-Database-1-1-2023.xlsx", col_names = TRUE, col_types = types )
sat.da.agg.a.dataUCS$'Lunch Mass kg(.)' <- vector("character", length = nrow(sat.da.agg.a.dataUCS))
sat.da.agg.a.dataUCS$`Lunch Mass kg(.)` <- data_prec$`Launch Mass (kg.)`[id1]
```


```{r}
# sistemo i nomi delle colonne in modo che siano uguali ( sia in nome che come variabili considerati)
names(dataUCS) <- c("Satellite Name", "Country", "Users", "Purpose", "Orbit", "Longitude of GEO","Perigee", "Apogee", "Eccentricity", "Inclination", "Period", "Mass", "Launch Date","Expected Lifetime",  "Final Date", "Status", "Effective Lifetime")
names(sat.da.agg.a.dataUCS) <- c("Satellite Name" ,    "Country"  ,          "Users"   ,           "Purpose"   ,         "Orbit"     ,        
 "Longitude of GEO"  , "Perigee"  ,          "Apogee"    ,         "Eccentricity"  ,     "Inclination"  ,     
"Period"      ,       "Mass"        ,       "Launch Date"   ,     "Status"        ,     "Final Date"      ,  
"Effective Lifetime" ,"Expected Lifetime" , "Continent"     ,     "Mass" ) 

sat.da.agg.a.dataUCS<-sat.da.agg.a.dataUCS[,-18] # Rimuovo continent
sat.da.agg.a.dataUCS<-sat.da.agg.a.dataUCS[,-12] # Rimuovo dry mass

dataUCS<- dataUCS[, names(sat.da.agg.a.dataUCS)]
dataUCS$Apogee <- as.numeric(dataUCS$Apogee)
sat.da.agg.a.dataUCS$`Final Date`<-as.character(sat.da.agg.a.dataUCS$`Final Date`)
dataUCS$`Effective Lifetime`<-as.numeric(dataUCS$`Effective Lifetime`)
```

Unisco i datasets
```{r}
library(dplyr)

#concateno 
joinData <- bind_rows(sat.da.agg.a.dataUCS,dataUCS)
table(joinData$Status)

# Caricare la libreria
library(openxlsx)
path = "C:/Users/alessandro/Documents/GitHub/NPS_project/Src"

# Scrivere il dataframe nel file Excel
write.xlsx(joinData,path, rowNames = FALSE)
```


AGGIUNGO COLONNA "CONTINENT" 

```{r}

library(readxl)
data <- read_excel("../Dataset/JoinDatasets.xlsx", col_names = TRUE)

paesi <- data$Country

# Carica il pacchetto
#install.packages("countrycode")

library(countrycode)
# Crea un dataframe di esempio con i paesi
dataframe_paesi <- data.frame(paese = paesi)
esa <- which(paesi == 'ESA')
multinat <- which(paesi == 'Multinational')

# Usa la funzione countrycode per ottenere il continente
continente <- countrycode(sourcevar = dataframe_paesi$paese, origin = "country.name", destination = "continent")
continente[esa] = 'Europe'
continente[multinat] = 'Multinational'

data <- cbind(data,Continent = continente)
save(data, file ="Data_new.Rdata")

install.packages("openxlsx")
library(openxlsx)


# Specifica il percorso del file Excel in cui desideri salvare il dataset
percorso_file <- "percorso/del/tuo/file/nomefile.xlsx"

# Salva il dataset in Excel
write.xlsx(data, file = 'JoinDatasets.xlsx', rowNames = FALSE)

```



Pulizia join:
```{r}
# per lettura dataset se prendi file da github
library(readxl)
data <- read_excel("../Dataset/JoinDatasets.xlsx", col_names = TRUE)
```


Summary:
```{r}
summary(data)
```




Sistemazione conflitti Country:

China/France -> China
India/France -> India
Japan/Singapore -> Japan
Poland/UK -> Poland
```{r}
data$Country[which(data$Country == "China/France")] <- "China"
data$Country[which(data$Country == "India/France")] <- "India"
data$Country[which(data$Country == "Japan/Singapore")] <- "Japan"
data$Country[which(data$Country == "Poland/UK")] <- "Poland"
```

Assegnazione Continent a paesi con continent NA, sono India Japan China -> Asia
```{r}
data$Country[which(is.na(data$Continent))]
data$Continent[which(is.na(data$Continent))] <- "Asia"
```


Sistemazione conflitti Users:

Civil/Government -> Civil
Commercial/Civil -> Commercial
Government/Civil -> Government
Government/Commercial -> Government
Military/Civil -> Military
```{r}
data$Users[which(data$Users == "Civil/Government")] <- "Civil"
data$Users[which(data$Users == "Commercial/Civil")] <- "Commercial"
data$Users[which(data$Users == "Government/Civil")] <- "Government"
data$Users[which(data$Users == "Government/Commercial")] <- "Government"
data$Users[which(data$Users == "Military/Civil")] <- "Military"
```


Sistemazione conflitti Purpose:

Communications/Technology Development -> Communications
Earth Observation/Technology Development -> Earth Observation
Earth/Space Observation -> Earth Observation
```{r}
data$Purpose[which(data$Purpose == "Communications/Technology Development")] <- "Communications"
data$Purpose[which(data$Purpose == "Earth Observation/Technology Development")] <- "Earth Observation"
data$Purpose[which(data$Purpose == "Earth/Space Observation")] <- "Earth Observation"
```



Rendi colonne factor:
```{r}
data$Country <- factor(data$Country, ordered = F)
print('Country')
table(data$Country)

data$Users <- factor(data$Users, ordered = F)
print('Users')
table(data$Users)

data$Purpose <- factor(data$Purpose, ordered = F)
print('Purpose')
table(data$Purpose)

data$Orbit <- factor(data$Orbit, ordered = F)
print('Orbit')
table(data$Orbit)

data$Status <- factor(data$Status, ordered = F)
print('Status')
table(data$Status)

data$Continent <- factor(data$Continent, ordered = F)
print('Continent')
table(data$Continent)
```

Numero di NA in ogni colonna
```{r}
df <- data.frame()

for (colonna in colnames(data)) {
    conteggio_na <- sum(is.na(data[[colonna]]))
    df <- rbind(df, data.frame(Colonna = colonna, Numero_di_NA = conteggio_na))
}

df
```

Rimuovi osservazioni con NA in periodo, launch date, mass:
```{r}
cols <- c("Period", "Launch Date", "Mass")

data <- data[complete.cases(data[, cols]), ]
```


Rendi POSIXct Final date:
```{r}
class(data$`Final Date`)

data$`Final Date` <- as.Date(data$`Final Date`)
data$`Final Date` <- as.POSIXct(data$`Final Date`)

class(data$`Final Date`)
```


Calcolo effective lifetime:
```{r}
data$`Effective Lifetime` <- round(as.numeric(data$`Final Date` - data$`Launch Date`)/365, digit = 2)

# Eliminare satelliti con tempi di vita effettivi negativi
data<- data[-which(data$`Effective Lifetime` < 0),]
```


Numero di NA in ogni colonna: rimangono solo 218 satelliti senza Expected Lifetime
```{r}
df <- data.frame()

for (colonna in colnames(data)) {
    conteggio_na <- sum(is.na(data[[colonna]]))
    df <- rbind(df, data.frame(Colonna = colonna, Numero_di_NA = conteggio_na))
}

df
```



AGGLOMERAMENTO CATEGORIE CON 1 OSSERVAZIONE 

```{r}
# VARIABILE PURPOSE
table(data$Purpose)

#tolgo l'osservazione con purpose unknown
id=which(data$Purpose=="Unknown")
data=data[-id,]
data$Purpose <- droplevels(data$Purpose)

# aggrego "Earth Science", "Educational" e "Surveillance" in "Earth Observation"
id=which(data$Purpose=="Earth Science")
data$Purpose[id]="Earth Observation"
id=which(data$Purpose=="Educational")
data$Purpose[id]="Earth Observation"
id=which(data$Purpose=="Surveillance")
data$Purpose[id]="Earth Observation"
data$Purpose <- droplevels(data$Purpose)

#aggrego "Technology Demonstration" in "Technology Development"
id=which(data$Purpose=="Technology Demonstration")
data$Purpose[id]="Technology Development"
data$Purpose <- droplevels(data$Purpose)

#aggrego "Space Observation" in "Space Science"
id=which(data$Purpose=="Space Observation")
data$Purpose[id]="Space Science"
data$Purpose <- droplevels(data$Purpose)


table(data$Purpose)


#VARIABILE CONTINENT
table(data$Continent)

```

Salvo dataset finale
```{r}
# Salva il dataset in Excel
library(openxlsx)

write.xlsx(data, file = 'JoinDatasets.xlsx', rowNames = FALSE)
```
