==> Src/02 - Test Set e Training Set.Rmd <==
```{r}
library(readxl)
library(dplyr)
library(caret)
library(openxlsx)

set.seed(123)


```


```{r}
data <- read_excel("../Dataset/JoinDatasets.xlsx", col_names = TRUE)
```

prendo come training test le osservazioni senza expected life time

```{r}
idna <- which(is.na(data$`Expected Lifetime`))
training_set0 <- data[idna,]
dim(training_set0) #217 obs
```

Ci aggiungo altre osservazioni per arrivare circa all'80% del dataset

```{r}
sub <- data[-idna,] # dataset con expected lifetime con 497 obs
index <- createDataPartition(sub$`Effective Lifetime`, p = 0.715, list = FALSE)
```

Crea il training set utilizzando l'indice e concateno
```{r}
training_set1 <- sub[index,]
training_set <- bind_rows(training_set0, training_set1)
dim(training_set)
test_set <- sub[-index, ]
dim(test_set)

```


```{r}
write.xlsx(test_set, file = 'TestSet.xlsx', rowNames = FALSE)
write.xlsx(training_set, file = 'TrainingSet.xlsx', rowNames = FALSE)

```
