==> Src/03 - Dataset inspection.Rmd <==
---
title: "NPS Project Survival Analysis"
author: "Simone Ciardulli"
date: "`r Sys.Date()`"
output: html_document
---


```{r}
# per lettura dataset se prendi file da github
library(readxl)
data <- read_excel("../Dataset/JoinDatasets.xlsx", col_names = TRUE)
i <- which(data$Purpose == 'Navigation' | data$Orbit == 'Elliptical' | data$Continent == 'Africa' | data$Continent == 'Multinational' | data$Continent == 'Oceania' )
data <- data[-i,]
```

Libraries:
```{r,'error=FALSE', 'warning=FALSE'}
library(readxl)
library(survival)
library(survminer)
library(dplyr) 
library(ggplot2)
library(knitr)
library(broom)
library(tidyr)
```

Summary:
```{r}
summary(data)
```

Rendi colonne factor:
```{r}
data$Country <- factor(data$Country, ordered = F)
print('Country')
table(data$Country)

data$Continent <- factor(data$Continent, ordered = F)
print('Continent')
table(data$Continent)

data$Users <- factor(data$Users, ordered = F)
print('Users')
table(data$Users)

data$Purpose <- factor(data$Purpose, ordered = F)
print('Purpose')
table(data$Purpose)

data$Orbit <- factor(data$Orbit, ordered = F)
print('Class of Orbit')
table(data$Orbit)

data$Status <- factor(data$Status, ordered = F)
print('Status')
table(data$Status)
```


Mischia righe
```{r}
set.seed(42)
data <- data[sample(nrow(data)), ]
```


Primo plot con tempi:
```{r}
data$ID <- factor(seq(1:nrow(data))) 
data_subs <- head(data)
time <- data_subs$'Effective Lifetime'

ggplot(data=data_subs,aes(x=ID,y=time)) + 
  geom_bar(stat='identity',width=0.2) +
  geom_point(aes(color=Status,shape=Status),size=6) +
  coord_flip()
```


```{r}
head(Surv(time=data$'Effective Lifetime',  # T_i
          event=data$Status == 'retired') # tell the object which rows have the event
     )
```

KM Estimator
```{r}
time <- data$'Effective Lifetime'
fit <- survfit(Surv(time,Status == 'retired') ~ 1, data = data)
```

Summary:
```{r}
summary(fit)
```

```{r}
kable(head(tidy(fit),20))
```

Median:
```{r}
surv_median(fit) #6.32
```


Plot of KM estimator curve
```{r}
ggsurvplot(fit, data = data,
           risk.table = TRUE, # Add risk table
           risk.table.col = "strata", # Change risk table color by groups
           surv.median.line = "hv", # Specify median survival
           ggtheme = theme_bw(), # Change ggplot2 theme
           break.time.by=5,
           title="Kaplan-Meier Curve for Satellites Survival")
```
NB: Per noi morte significa che il satellite è ancora integro ma non in funzione. 



Cumulative Failure Probability (CFP):

$$CFP(t) = P(T\leq t)$$ 
so it can be estimated as:
$$1-S(t)$$

```{r}
cumulative_incidence <- 1 - fit$surv
```

CFP plot: (ggsurvplot() function specifying the option fun=‘event’)
```{r}
ggsurvplot(fit, data = data,
           risk.table = TRUE, # Add risk table
           risk.table.col = "strata", # Change risk table color by groups
           surv.median.line = "hv", # Specify median survival
           ggtheme = theme_bw(), # Change ggplot2 theme
           break.time.by=5,
           fun='event',
           title="Cumulative Incidence Curve")
```

Cumulative Hazard
```{r}
ggsurvplot(fit, data = data,
           risk.table = TRUE, # Add risk table
           ggtheme = theme_bw(), # Change ggplot2 theme
           cumhaz.median.line = "hv", # Specify median survival
           break.time.by=5,
           fun='cumhaz',
           title="Cumulative Hazard Curve for Satellites Survival")
```



1) We construct one curve for each Continent: -> Difference is not statistically significant
```{r}
fit.Continent <- survfit(Surv(time,Status == 'retired') ~ Continent, data=data)
ggsurvplot(fit.Continent, data = data,
           risk.table = TRUE, # Add risk table
           risk.table.col = "strata",
           ggtheme = theme_bw(), # Change ggplot2 theme
           break.time.by=5,
           title="Kaplan-Meier Curve for Satellites Survival by Continent")
```

Log-rank test:
```{r}
log_rank_test <- survdiff(Surv(time, Status == 'retired') ~ Continent, data=data)
log_rank_test
```

Hazard ratio: 
```{r}
hazard_ratio <- (log_rank_test$obs[1]/log_rank_test$exp[1])/(log_rank_test$obs[2]/log_rank_test$exp[2])
hazard_ratio
hazard_ratio <- (log_rank_test$obs[2]/log_rank_test$exp[2])/(log_rank_test$obs[1]/log_rank_test$exp[1])
hazard_ratio
```



2) We construct one curve for each User: 
```{r}
fit.Users <- survfit(Surv(time, Status == 'retired') ~ Users, data=data)

ggsurvplot(fit.Users, data = data,
           risk.table = TRUE, # Add risk table
           risk.table.col = "strata",
           ggtheme = theme_bw(), # Change ggplot2 theme
           break.time.by=5,
           title="Kaplan-Meier Curve for Satellites Survival by Users")
```

Log-rank test:

```{r}
log_rank_test <- survdiff(Surv(time, Status == 'retired') ~ Users, data=data)
log_rank_test # p = <2e-16
```

Hazard ratio:
```{r}
hazard_ratio <- (log_rank_test$obs[1]/log_rank_test$exp[1])/(log_rank_test$obs[2]/log_rank_test$exp[2])
hazard_ratio 

hazard_ratio <- (log_rank_test$obs[2]/log_rank_test$exp[2])/(log_rank_test$obs[3]/log_rank_test$exp[3])
hazard_ratio

hazard_ratio<- (log_rank_test$obs[3]/log_rank_test$exp[3])/(log_rank_test$obs[4]/log_rank_test$exp[4])
hazard_ratio
```

Factors ordered by risk (da più rischioso): Civil, Commercial, Miitary, Government 


3) We construct one curve for each Purpose -> Difference is statistically significant
```{r}
fit.Purpose <- survfit(Surv(time, Status == 'retired') ~ Purpose, data=data)

ggsurvplot(fit.Purpose, data = data,
           risk.table = TRUE, # Add risk table
           risk.table.col = "strata",
           ggtheme = theme_bw(), # Change ggplot2 theme
           break.time.by=5,
           title="Kaplan-Meier Curve for Satellites Survival by Purpose")
```

Log-rank test:

```{r}
log_rank_test <- survdiff(Surv(time, Status == 'retired') ~ Purpose, data=data)
log_rank_test # p < 2e-16 
```

Hazard ratio: 
```{r}
hazard_ratio <- (log_rank_test$obs[1]/log_rank_test$exp[1])/(log_rank_test$obs[2]/log_rank_test$exp[2])
hazard_ratio 

hazard_ratio <- (log_rank_test$obs[2]/log_rank_test$exp[2])/(log_rank_test$obs[3]/log_rank_test$exp[3])
hazard_ratio

hazard_ratio<- (log_rank_test$obs[3]/log_rank_test$exp[3])/(log_rank_test$obs[4]/log_rank_test$exp[4])
hazard_ratio
```

Factors ordered by risk (dal più rischioso): Technology Development, Space Science,  Earth Observation,  Communications  


4) We construct one curve for Orbit:
```{r}
fit.Orbit <- survfit(Surv(time,data$Status == 'retired') ~ Orbit, data=data)

ggsurvplot(fit.Orbit, data = data,
           risk.table = TRUE, # Add risk table
           risk.table.col = "strata",
           ggtheme = theme_bw(), # Change ggplot2 theme
           break.time.by=5,
           title="Kaplan-Meier Curve for Satellites Survival by Orbit")
```


Log-rank test:

```{r}
log_rank_test <- survdiff(Surv(time, Status == 'retired') ~ Orbit, data=data)
log_rank_test # p < 2e-16
```

Hazard ratio
```{r}
hazard_ratio <- (log_rank_test$obs[1]/log_rank_test$exp[1])/(log_rank_test$obs[2]/log_rank_test$exp[2])
hazard_ratio 

```


Factors ordered by risk: LEO, GEO
