==> Src/05 -  RP Intervals.Rmd <==
```{r}
library(readxl)
data <- read_excel("../dataset/TrainingSet.xlsx", col_names = TRUE)
```

Libraries:
```{r,'error=FALSE', 'warning=FALSE'}
library(readxl)
library(survival)
library(survminer)
library(dplyr) 
library(ggplot2)
library(knitr)
library(broom)
library(tidyr)
```

Summary:
```{r}
summary(data)
```
Conto numero eventi per ogni covariata
```{r}
table(data$Users, data$Status)
table(data$Purpose, data$Status)
table(data$Orbit, data$Status)
table(data$Continent, data$Status)


```
Dalla letteratura risulta che non ho abbastanza eventi per Navigation, Elliptical, Africa, Multinational, Oceania
Simulation work has suggested that at least 10 events need to be observed for each covariate considered, and anything less will lead to problems.

Rimuoviamo queste classi
```{r}
i <- which(data$Purpose == 'Navigation' | data$Orbit == 'Elliptical' | data$Continent == 'Africa' | data$Continent == 'Multinational' | data$Continent == 'Oceania' )
data <- data[-i,]
```

A livello di continenti è un confronto tra America, Asia e Europa: Maggiori potenze economiche 

Controlliamo
```{r}
table(data$Users, data$Status)
table(data$Purpose, data$Status)
table(data$Orbit, data$Status)
table(data$Continent, data$Status)
table(data$Status)
```


Trasformo Eccentricità per avere un miglior grafico dei residui
```{r}
i <- which(data$Eccentricity == 0)
data$Eccentricity[i] = 1
data$log.Eccentricity <- -log(data$Eccentricity)
```


```{r}
time <- data$`Effective Lifetime`
status <- data$Status

covariates <- c( "Perigee" ,"Apogee", 'log.Eccentricity', "Inclination", "Period", "Mass", "Users", "Purpose", "Continent", "Orbit")

cox1bis <- coxph(Surv(time,status == 'retired') ~ ., data = data[,covariates])
summary(cox1bis)

```


Dimostro che perigeo non è significant con RP intervals

```{r}
n <- nrow(data)
perigee_coef <- cox1bis$coefficients[1]
B <- 1000
T.perigee = numeric(B)
set.seed(2022)
for(b in 1:B) {
  boot_id <- sample(x = 1:n,size = n,replace = TRUE)
  df_boot <- data[boot_id,]
  fit_cox_boot <- coxph(Surv(time,status == 'retired') ~ ., data = df_boot[,covariates])
  T.perigee[b] = fit_cox_boot$coefficients[1]
}


alpha <- 0.05
right.quantile <- quantile(T.perigee, 1 - alpha/2)
left.quantile <- quantile(T.perigee, alpha/2)
CI.RP.perigee <-
  c(
    perigee_coef - (right.quantile - perigee_coef),
    perigee_coef,
    perigee_coef - (left.quantile- perigee_coef))
names(CI.RP.perigee)=c('lwr','pointwise','upr')
CI.RP.perigee

```
The coef is not significant at 5% ( zero incluso) 
Rimuovo Perigee
```{r}
time <- data$`Effective Lifetime`
status <- data$Status

covariates <- c("Apogee", 'log.Eccentricity', "Inclination", "Period", "Mass", "Users", "Purpose", "Continent", "Orbit")

cox2bis <- coxph(Surv(time,status == 'retired') ~ ., data = data[,covariates])
summary(cox2bis)

```

Strata su purpose
```{r}
time <- data$`Effective Lifetime`
status <- data$Status

cox2bis <- coxph(Surv(time,status == 'retired') ~  Apogee + log.Eccentricity + Inclination + Period +
                Mass + Users + strata(Purpose) + Continent + Orbit, data = data)
summary(cox2bis)

```

Dimostro che bisgona rimuovere Inclination

```{r}
n <- nrow(data)
inc_coef <- cox2bis$coefficients[3]
B <- 1000
T.Inc = numeric(B)
set.seed(2022)
for(b in 1:B) {
  boot_id <- sample(x = 1:n,size = n,replace = TRUE)
  df_boot <- data[boot_id,]
  fit_cox_boot <- coxph(Surv(time,status == 'retired') ~  Apogee + log.Eccentricity + Inclination + Period +
                Mass + Users + strata(Purpose) + Continent + Orbit, data = df_boot)
  T.Inc[b] = fit_cox_boot$coefficients[3]
}


alpha <- 0.05
right.quantile <- quantile(T.Inc, 1 - alpha/2)
left.quantile <- quantile(T.Inc, alpha/2)
CI.RP.Inc <-
  c(
    inc_coef - (right.quantile - inc_coef),
    inc_coef,
    inc_coef - (left.quantile- inc_coef))
names(CI.RP.perigee)=c('lwr','pointwise','upr')
CI.RP.Inc

```
The coef is not significant at 5% ( zero incluso)


Rimuovo Inclination
```{r}
time <- data$`Effective Lifetime`
status <- data$Status

cox3bis <- coxph(Surv(time,status == 'retired') ~  Apogee + log.Eccentricity  + Period +
                Mass + Users + strata(Purpose) + Continent + Orbit, data = data)
summary(cox3bis)

```

Strata su orbit

```{r}
time <- data$`Effective Lifetime`
status <- data$Status

cox4bis <- coxph(Surv(time,status == 'retired') ~  Apogee + log.Eccentricity  + Period +
                Mass + Users + strata(Purpose) + Continent + strata(Orbit), data = data)
summary(cox4bis)

```

Dimostro che ha senso rimuovere Periodo

```{r}
n <- nrow(data)
per_coef <- cox4bis$coefficients[3]
B <- 1000
T.Periodo = numeric(B)
set.seed(2022)
for(b in 1:B) {
  boot_id <- sample(x = 1:n,size = n,replace = TRUE)
  df_boot <- data[boot_id,]
  fit_cox_boot <- coxph(Surv(time,status == 'retired') ~  Apogee + log.Eccentricity  + Period +
                Mass + Users + strata(Purpose) + Continent + strata(Orbit), data = df_boot)
  T.Periodo[b] = fit_cox_boot$coefficients[3]
}


alpha <- 0.05
right.quantile <- quantile(T.Periodo, 1 - alpha/2)
left.quantile <- quantile(T.Periodo, alpha/2)
CI.RP.Periodo <-
  c(
    per_coef - (right.quantile - per_coef),
    per_coef,
    per_coef - (left.quantile- per_coef))
names(CI.RP.perigee)=c('lwr','pointwise','upr')
CI.RP.Periodo

```
The coef is not significant at a confidecen level of 5%.


Rimuovo Periodo

```{r}
time <- data$`Effective Lifetime`
status <- data$Status

cox5bis <- coxph(Surv(time,status == 'retired') ~  Apogee + log.Eccentricity   +
                Mass + Users + strata(Purpose) + Continent + strata(Orbit), data = data)
summary(cox5bis)

```
