==> Src/07 - Test.Rmd <==
---
title: "test&anova"
output: html_document
date: "2023-11-28"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}

library("openxlsx")
library(readxl)
data<-  read_excel("../Dataset/TestSet.xlsx", col_names = TRUE)
#predicted_lifetime = read.table("predicted_lifetime.txt", header = TRUE)
predicted_lifetime = read.table("../Dataset/pred2.txt", header = TRUE)



```

```{r}
library(readxl)
library(survival)
library(survminer)
library(dplyr) 
library(ggplot2)
library(knitr)
library(broom)
library(tidyr)
```

modifico il dataset come in "Cox Model - Assunzioni Verificate.Rmd"

```{r}
i <- which(data$Purpose == 'Navigation' | data$Orbit == 'Elliptical' | data$Continent == 'Africa' | data$Continent == 'Multinational' | data$Continent == 'Oceania' )
data <- data[-i,]
```

```{r}
table(data$Users, data$Status)
table(data$Purpose, data$Status)
table(data$Orbit, data$Status)
table(data$Continent, data$Status)
table(data$Status)
```

```{r}
i <- which(data$Eccentricity == 0)
data$Eccentricity[i] = 1
data$log.Eccentricity <- -log(data$Eccentricity)
```

riguardo le table a dataset ultimato

```{r}
table(data$Users, data$Status)
table(data$Purpose, data$Status)
table(data$Orbit, data$Status)
table(data$Continent, data$Status)
table(data$Status)
```

I civili sono solo 2 quindi li incorporo con qualcun altro. Space science e technology development sempre pochi==\>potrei incorporare space science con earth observation e technology development con communications?

PERMUTATIONAL ONE-WAY ANOVA PRED - EXP-LIFETIME -\> CONTINENT

```{r}
B = 1000
predlife = predicted_lifetime$exp.dur
explife = data$`Expected Lifetime`
diff = predlife-explife
continent = as.factor(data$Continent)

g = nlevels(continent)
n = dim(data)[1]

plot(continent, diff, xlab='continent',col=rainbow(g),main='Original Data')


fit_c <- aov(diff ~ continent)
summary(fit_c)

T0 <- summary(fit_c)[[1]][1,4]  # extract the test statistic
T0


T_stat <- numeric(B) 

for(perm in 1:B){
  # Permutation:
  permutation <- sample(1:n)
  diff_perm <- diff[permutation]
  fit_perm <- aov(diff_perm ~ continent)
  
  # Test statistic:
  T_stat[perm] <- summary(fit_perm)[[1]][1,4]
}


hist(T_stat,xlim=range(c(T_stat,T0)),breaks=30)
abline(v=T0,col=3,lwd=2)

plot(ecdf(T_stat),xlim=c(-1,20))
abline(v=T0,col=3,lwd=4)

# p-value
p_val <- sum(T_stat>=T0)/B
p_val #0.038
```

Il continente è un fattore significativo ==\>c'è evidenza statistica per permetterci di affermare che alcuni continenti mentono in media più di altri

PERMUTATIONAL ONE-WAY ANOVA predicted- expected -\> PURPOSE

```{r}

purpose = as.factor(data$Purpose)
g = nlevels(purpose)
n = dim(data)[1]

plot(purpose, diff, xlab='purpose',col=rainbow(g),main='Original Data')


fit_p <- aov(diff ~ purpose)
summary(fit_p)

T0 <- summary(fit_p)[[1]][1,4]  # extract the test statistic
T0


T_stat <- numeric(B) 

for(perm in 1:B){
  # Permutation:
  permutation <- sample(1:n)
  diff_perm <- diff[permutation]
  fit_perm <- aov(diff_perm ~ purpose)
  
  # Test statistic:
  T_stat[perm] <- summary(fit_perm)[[1]][1,4]
}


hist(T_stat,xlim=range(c(T_stat,T0)),breaks=30)
abline(v=T0,col=3,lwd=2)

plot(ecdf(T_stat),xlim=c(-1,80))
abline(v=T0,col=3,lwd=4)

# p-value
p_val <- sum(T_stat>=T0)/B
p_val #0

```

Like continents: there is evidence but we have to pay attention on the dimension of the sample

PERMUTATIONAL ONE-WAY ANOVA EXP-LIFETIME -\> USERS

```{r}

users = as.factor(data$Users)
g = nlevels(users)
n = dim(data)[1]

plot(users, diff, xlab='users',col=rainbow(g),main='Original Data')


fit_u <- aov(diff ~ users)
summary(fit_u)

T0 <- summary(fit_u)[[1]][1,4]  # extract the test statistic
T0


T_stat <- numeric(B) 

for(perm in 1:B){
  # Permutation:
  permutation <- sample(1:n)
  diff_perm <- diff[permutation]
  fit_perm <- aov(diff_perm ~ users)
  
  # Test statistic:
  T_stat[perm] <- summary(fit_perm)[[1]][1,4]
}


hist(T_stat,xlim=range(c(T_stat,T0)),breaks=30)
abline(v=T0,col=3,lwd=2)

plot(ecdf(T_stat),xlim=c(-1,80))
abline(v=T0,col=3,lwd=4)

# p-value
p_val <- sum(T_stat>=T0)/B
p_val #0.332

```

PERMUTATIONAL TWO-WAYS ANOVA PREDLIFETIME - EXPLIFETIME -\> CONTINENT-PURPOSE

```{r}
summary.aov(aov(diff ~ continent + purpose + continent:purpose)) 
T0_cp <- summary.aov(aov(diff ~ continent + purpose + continent:purpose))[[1]][3,4]
T0_cp

aov.H0cp <- aov(diff ~ continent + purpose) #reduced model 
aov.H0cp
residuals.H0cp <- aov.H0cp$residuals


T_cp<- numeric(B)
for(perm in 1:B){
  permutation <- sample(n)
  residuals.H0cp <- residuals.H0cp[permutation]
  diff.perm.H0cp <- aov.H0cp$fitted + residuals.H0cp
  T_cp[perm] <- summary.aov(aov(diff.perm.H0cp ~ continent + purpose + continent:purpose))[[1]][3,4]
}

sum(T_cp >= T0_cp)/B #0
```

The interaction between continent and purpose is SIGNIFICANT

PERMUTATIONAL TWO-WAYS ANOVA EXP-LIFETIME -\> USERS-PURPOSE

```{r}
summary.aov(aov(diff ~ users + purpose + users:purpose)) 
T0_cp <- summary.aov(aov(diff ~ users + purpose + users:purpose))[[1]][3,4]
T0_cp

aov.H0cp <- aov(diff ~ users + purpose) #reduced model 
aov.H0cp
residuals.H0cp <- aov.H0cp$residuals


T_cp<- numeric(B)
for(perm in 1:B){
  permutation <- sample(n)
  residuals.H0cp <- residuals.H0cp[permutation]
  diff.perm.H0cp <- aov.H0cp$fitted + residuals.H0cp
  T_cp[perm] <- summary.aov(aov(diff.perm.H0cp ~ users + purpose + users:purpose))[[1]][3,4]
}

sum(T_cp >= T0_cp)/B #0.055
```

The interaction between users and purpose is not significant

------------------------------------------------------------------------

TEST FOR THE DISTRIBUTION OF THE DIFFERENCE between predicted and expected lifetime

SIAMO NEL CASO DI PAIRED DATA ==\> sia noi che le agenzie spaziali misuriamo la stessa quantità: il tempo di vita atteso

```{r}
p.value <- wilcox.test(explife,y=predlife, paired=T, conf.level = 1-alpha)$p.value
p.value #0.0004405377
```

PERMUTATION TEST

Independence assumption is not verified !!!

H0: X and Y have the same distribution vs H1: X and Y have different distributions

```{r}
perm_t_test=function(x,y,iter=1e3){ #conditional MC -> n.iterations
  
  T0=abs(mean(x)-mean(y))  # define the test statistic
  T_stat=numeric(iter) # a vector to store the values of each iteration
  x_pooled=c(x,y) # pooled sample
  n=length(x_pooled)
  n1=length(x)
  
  for(perm in 1:iter){ # loop for conditional MC
    # permutation:
    permutation <- sample(1:n)
    x_perm <- x_pooled[permutation]
    x1_perm <- x_perm[1:n1]
    x2_perm <- x_perm[(n1+1):n]
    # test statistic:
    T_stat[perm] <- abs(mean(x1_perm) - mean(x2_perm))
    
  }
  
  # p-value
  p_val <- sum(T_stat>=T0)/iter
  return(p_val)
}


p.value <- perm_t_test(explife, predlife)
p.value #0

```

TWO SAMPLE PAIRED UNIVARIATE PERMUTATION TEST (the one that we consider)

H0 : X and Y have the same distribution

```{r}
t1=explife
t2=predlife
delta.0 <- 0
diff <- t1-t2
n=length(diff)

T20 <- abs(mean(diff-delta.0))

T2 <- numeric(B)
for(perm in 1:B)
  {
  # Random permutation
  # obs: exchanging data within couples means changing the sign of the difference
  signs.perm <- rbinom(n, 1, 0.5)*2 - 1
  
  diff_perm <- diff * signs.perm
  diff.mean_perm <- mean(diff_perm)
  
  T2[perm] <- abs(diff.mean_perm-delta.0)
}

hist(T2,xlim=range(c(T2,T20)),breaks=100)
abline(v=T20,col=3,lwd=4)

plot(ecdf(T2))
abline(v=T20,col=3,lwd=4)

p_val <- sum(T2>=T20)/B
p_val #0
```

ALTERNATIVE:

Instead of performing inference on the distribution of the difference, we treat them "separately". We have 2\^N different permutations, since the **exchangeability is only within the pairs**. We choose as test statistic the norm of the differences of the means of both samples.

```{r}
T20 <- norm(as.matrix(mean(t1) - mean(t2)))

T2 <- numeric(B)
t.full <- rbind(t1, t2)
for(perm in 1:B)
  {
  # Random permutation
  # N.B. exchangeability is only within pairs
  perm.indices.t1 <- seq(1, n) + n * rbinom(n,1, 0.5)
  t1.perm <- t.full[perm.indices.t1]
  t2.perm <- t.full[-perm.indices.t1]
  
  T2[perm] <- norm(as.matrix(((mean(t1.perm)) - mean(t2.perm))))
}

hist(T2,xlim=range(c(T2,T20)),breaks=100)
abline(v=T20,col=3,lwd=4)

plot(ecdf(T2))
abline(v=T20,col=3,lwd=4)

# p-value
p_val <- sum(T2>=T20)/B
p_val #0
```

------------------------------------------------------------------------

DDPLOT to compare the 2 populations

```{r}
windows()

DepthProc::ddPlot(x = as.matrix(predlife), # use as.matrix since ddplot usually receives multivariate data
                  y = as.matrix(explife),depth_params = list(method='Tukey'))
```

------------------------------------------------------------------------

PERMUTATIONAL CONFIDENCE INTERVALS

Let's use the two sample paired univariate permutation test

```{r}

t1=explife
t2=predlife
diff <- t2-t1

test1 <- function(diff, delta.0, B = 1000) {
  
  data_trans <- diff - delta.0
  T0 <- abs(mean(data_trans)) # statistic T=abs(mean(diff-delta.0))  (univariate case) 
  T_perm <- numeric(B)
  n1 <- length(diff)
  p <-1 
  

  for(perm in 1:B)
    {
    # Random permutation
    # obs: exchanging data within couples means changing the sign of the difference
    signs.perm <- rbinom(n1, 1, 0.5)*2 - 1
    data_trans_perm <- data_trans * signs.perm
    T_perm[perm] <- abs(mean(data_trans_perm)) 
    
  }
  
  p_val <- sum(T_perm >= T0)/B
  
  
  return(p_val)

}


grid=seq(-5,1,by=0.1)
length(grid)

library(pbapply)
library(parallel)

n_cores <- detectCores()
n_cores

cl = makeCluster(n_cores)

clusterExport(cl,varlist=list("diff","test1"))

perm_wrapper <- function(grid_point) {
  test1(diff, grid_point, B=2000)  
}

pval_function <- pbsapply(grid, perm_wrapper, cl = cl)

# pval_function <- numeric(length(grid))
# i = 0
# for (theta in grid) {
#   pval_function[i] = test1(theta)
#   i = i+1
# }


alpha <- 0.05  # set the significance level
plot(grid, pval_function, type = "l")  # plot p-value function
values.within.CI <- grid[pval_function > alpha]
CI <- range(values.within.CI)  # obtain the confidence interval
abline(v=CI[1], col="red")
abline(v=CI[2], col="red")
abline(v = mean(diff), lty=2)


```

BOOTSTRAP DISTRIBUTION & CONFIDENCE INTERVALS for the mean of the difference

```{r}

library(pbapply)
library(parallel)

T.obs <- mean(diff)

cl=makeCluster(parallel::detectCores()/2)
clusterExport(cl=cl,list('diff'))

T.boot=pbreplicate(B,  
                   mean(sample(diff, replace = T)),
                   cl=cl)

plot(ecdf(T.boot), main='Sample mean of the difference')
abline(v = T.obs, lty=2)

```

Probability that the difference between the explife and the predlife (" amount of lie ") is \< x years wrt x years

ex: probability that explife-predlife \< -3 (i.e the explife is underestimated by more than 3 years) is about 0.005.

```{r}

var=var(T.boot)
bias=mean(T.boot)-T.obs
RMSE=sqrt(var+bias^2)

# REVERSE PERCENTILE INTERVALS
alpha <- 0.05

right.quantile <- quantile(T.boot, 1 - alpha/2)
left.quantile  <- quantile(T.boot, alpha/2)



CI.RP <- c(T.obs - (right.quantile - T.obs), T.obs - (left.quantile - T.obs))
names(CI.RP)=c('lwr','upr')

plot(ecdf(T.boot), main='Sample mean of the difference')
abline(v = T.obs, lty=2)
abline(v = CI.RP)


```

Bootstrap reverse percentile intervals are very similar to the permutation confidence intervals.

BOOTSTRAP CONFIDENCE INTERVALS FOR THE DIFFERENCE BETWEEN PREDICTED AND EXPECTED CONSIDERING ONLY THE CONTINENTS:

ASIA CI

```{r}
library(pbapply)
library(parallel)

id_Asia=which(data$Continent=="Asia")
diff1=diff[id_Asia]

T.obs1 <- mean(diff1)

cl=makeCluster(parallel::detectCores()/2)
clusterExport(cl=cl,list('diff1'))

T.boot1=pbreplicate(B,  
                   mean(sample(diff1, replace = T)),
                   cl=cl)

plot(ecdf(T.boot1), main='Sample mean of the difference')
abline(v = T.obs1, lty=2)

var=var(T.boot1)
bias=mean(T.boot1)-T.obs1
RMSE=sqrt(var+bias^2)

# REVERSE PERCENTILE INTERVALS
alpha <- 0.05

right.quantile <- quantile(T.boot1, 1 - alpha/2)
left.quantile  <- quantile(T.boot1, alpha/2)



CI.RP1 <- c(T.obs1 - (right.quantile - T.obs1), T.obs1 - (left.quantile - T.obs1))
names(CI.RP1)=c('lwr','upr')

plot(ecdf(T.boot1), main='Sample mean of the difference for Asia')
abline(v = T.obs1, lty=2)
abline(v = CI.RP1)
```

EUROPE CI

```{r}
library(pbapply)
library(parallel)

id_Europe=which(data$Continent=="Europe")
diff2=diff[id_Europe]

T.obs2 <- mean(diff2)

cl=makeCluster(parallel::detectCores()/2)
clusterExport(cl=cl,list('diff2'))

T.boot2=pbreplicate(B,  
                   mean(sample(diff2, replace = T)),
                   cl=cl)

plot(ecdf(T.boot2), main='Sample mean of the difference')
abline(v = T.obs2, lty=2)

var=var(T.boot2)
bias=mean(T.boot2)-T.obs2
RMSE=sqrt(var+bias^2)

# REVERSE PERCENTILE INTERVALS
alpha <- 0.05

right.quantile <- quantile(T.boot2, 1 - alpha/2)
left.quantile  <- quantile(T.boot2, alpha/2)



CI.RP2 <- c(T.obs2 - (right.quantile - T.obs2), T.obs2 - (left.quantile - T.obs2))
names(CI.RP2)=c('lwr','upr')

plot(ecdf(T.boot2), main='Sample mean of the difference for Europe')
abline(v = T.obs2, lty=2)
abline(v = CI.RP2)
```

AMERICA CI

```{r}
library(pbapply)
library(parallel)

id_America=which(data$Continent=="Americas")
diff3=diff[id_America]

T.obs3 <- mean(diff3)

cl=makeCluster(parallel::detectCores()/2)
clusterExport(cl=cl,list('diff3'))

T.boot3=pbreplicate(B,  
                   mean(sample(diff3, replace = T)),
                   cl=cl)

plot(ecdf(T.boot3), main='Sample mean of the difference')
abline(v = T.obs3, lty=2)

var=var(T.boot3)
bias=mean(T.boot3)-T.obs3
RMSE=sqrt(var+bias^2)

# REVERSE PERCENTILE INTERVALS
alpha <- 0.05

right.quantile <- quantile(T.boot3, 1 - alpha/2)
left.quantile  <- quantile(T.boot3, alpha/2)



CI.RP3 <- c(T.obs3 - (right.quantile - T.obs3), T.obs3 - (left.quantile - T.obs3))
names(CI.RP3)=c('lwr','upr')

plot(ecdf(T.boot3), main='Sample mean of the difference for Americas')
abline(v = T.obs3, lty=2)
abline(v = CI.RP3)
```

PLOT ALLINEATI

```{r}
windows() 
par(mfrow = c(3, 1))


plot(ecdf(T.boot1), main='Sample mean of the difference for Asia')
abline(v = T.obs1, lty=2)
abline(v = CI.RP1)

plot(ecdf(T.boot2), main='Sample mean of the difference for Europe')
abline(v = T.obs2, lty=2)
abline(v = CI.RP2)

plot(ecdf(T.boot3), main='Sample mean of the difference for Americas')
abline(v = T.obs3, lty=2)
abline(v = CI.RP3)
```

Notiamo che la differenza non è significativa per l'Asia, mentre lo è sia per l'Europa che per l'America.

Visto che il nostro scopo è quello di capire dove è meglio rivolgerci come startup, testiamo se la differenza tra Europa e America è significativa

```{r}
library(pbapply)
library(parallel)

diffEA=diff2-diff3 #Europe - America

T.obs <- mean(diffEA)

cl=makeCluster(parallel::detectCores()/2)
clusterExport(cl=cl,list('diffEA'))

T.boot=pbreplicate(B,  
                   mean(sample(diffEA, replace = T)),
                   cl=cl)

plot(ecdf(T.boot), main='Sample mean of the difference')
abline(v = T.obs, lty=2)

var=var(T.boot)
bias=mean(T.boot)-T.obs
RMSE=sqrt(var+bias^2)

# REVERSE PERCENTILE INTERVALS
alpha <- 0.05

right.quantile <- quantile(T.boot, 1 - alpha/2)
left.quantile  <- quantile(T.boot, alpha/2)



CI.RP <- c(T.obs - (right.quantile - T.obs), T.obs - (left.quantile - T.obs))
names(CI.RP)=c('lwr','upr')

#plot(ecdf(T.boot), main='Sample mean of the difference of our response between Europe and Americas')
#abline(v = T.obs, lty=2)
#abline(v = CI.RP)
```

```{r}
windows()
#par(mfrow = c(3, 1))

y=c(T.obs1, T.obs2, T.obs3)
x=c("Asia","Europe","Americas")
df = data.frame(x=x, y =y)
L=c(-1.543691,-7.7649421, -3.297262 )
U=c(5.836645, -0.5387716,-1.448506 )

CI_Americas <- c(-3.297262, -2.428359, -1.448506)
CI_Europe <- c(-7.7649421, -3.935845, -0.5387716)
CI_Asia <- c(-1.543691, 1.943106, 5.836645)

ggplot(df, aes(x = x, y = y)) +
    geom_errorbar(aes(ymax = U, ymin = L), width = 0.3) +
    geom_point(size = 4, col = "darkorange") +
    coord_flip() +
    labs(x = "Continents",
         y = "Disparity in Predicted vs. Expected Lifetime") 

```

\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\-\--

TEST E CI INIZIALI PER DIFFERENZA expected-effective

```{r}
id_ret= which(data$Status=="retired")
effectlife=data$`Effective Lifetime`
explife=data$`Expected Lifetime`
t1=effectlife[id_ret]
t2=explife[id_ret]

diff <- t1-t2
B = 1000
test1 <- function(diff, delta.0, B = 1000) {
  
  data_trans <- diff - delta.0
  T0 <- abs(mean(data_trans)) # statistic T=abs(mean(diff-delta.0))  (univariate case) 
  T_perm <- numeric(B)
  n1 <- length(diff)
  p <-1 
  

  for(perm in 1:B)
    {
    # Random permutation
    # obs: exchanging data within couples means changing the sign of the difference
    signs.perm <- rbinom(n1, 1, 0.5)*2 - 1
    data_trans_perm <- data_trans * signs.perm
    T_perm[perm] <- abs(mean(data_trans_perm)) 
    
  }
  
  p_val <- sum(T_perm >= T0)/B
  
  
  return(p_val)
}

grid=seq(-5,1,by=0.1)
length(grid)

library(pbapply)
library(parallel)

n_cores <- detectCores()
n_cores

cl = makeCluster(n_cores)

clusterExport(cl,varlist=list("diff","test1"))

perm_wrapper <- function(grid_point) {
  test1(diff, grid_point, B=2000)  
}

pval_function <- pbsapply(grid, perm_wrapper, cl = cl)

# pval_function <- numeric(length(grid))
# i = 0
# for (theta in grid) {
#   pval_function[i] = test1(theta)
#   i = i+1
# }


alpha <- 0.05  # set the significance level
plot(grid, pval_function, type = "l")  # plot p-value function
values.within.CI <- grid[pval_function > alpha]
CI <- range(values.within.CI)  # obtain the confidence interval
abline(v=CI[1], col="red")
abline(v=CI[2], col="red")
abline(v = mean(diff), lty=2)



library(pbapply)
library(parallel)

T.obs <- mean(diff)

cl=makeCluster(parallel::detectCores()/2)
clusterExport(cl=cl,list('diff'))

T.boot=pbreplicate(B,  
                   mean(sample(diff, replace = T)),
                   cl=cl)

plot(ecdf(T.boot), main='Sample mean of the difference')
abline(v = T.obs, lty=2)

var=var(T.boot)
bias=mean(T.boot)-T.obs
RMSE=sqrt(var+bias^2)

# REVERSE PERCENTILE INTERVALS
alpha <- 0.05

right.quantile <- quantile(T.boot, 1 - alpha/2)
left.quantile  <- quantile(T.boot, alpha/2)



CI.RP <- c(T.obs - (right.quantile - T.obs), T.obs - (left.quantile - T.obs))
names(CI.RP)=c('lwr','upr')

plot(ecdf(T.boot), main='Sample mean of the difference')
abline(v = T.obs, lty=2)
abline(v=CI.RP)
```

shapiro_test

```{r}
shapiro.test(diff)
```

```{r}
#DDPLOT
windows()
DepthProc::ddPlot(x = as.matrix(effectlife), # use as.matrix since ddplot usually receives multivariate data
                  y = as.matrix(explife),depth_params = list(method='Tukey'))


#PLOT BOOTSTRAP CI
windows()

CI=CI.RP

df <- data.frame(y = c(CI[1], CI[2]), label = c("Lower", "Upper"))

# Creazione del grafico con ggplot2
ggplot(df, aes(x = 1, y = y)) +
  geom_errorbar(aes(ymin = CI[1], ymax = CI[2]), width = 0.2, color = "blue", size = 1) +
  geom_point(color = "blue", size = 3) +
  geom_hline(yintercept = mean(CI), linetype = "dashed", color = "green", size = 1) +
  geom_hline(yintercept = 0, linetype = "dashed", color = "red", size = 1) +
  labs(title = "Intervallo di Confidenza", y = "Confidence Interval") +
  theme_minimal()
```
